[ARTIFY-23] Refactor package name
