--- model/source.go
@@ -1,4 +1,0 @@
-package model
-
-type Source struct {
-}
