+++ models/photo.go
@@ -7,7 +7,6 @@ import (
 	"github.com/gobuffalo/pop"
 	"github.com/gobuffalo/uuid"
 	"github.com/gobuffalo/validate"
-	"github.com/gobuffalo/validate/validators"
 )
 
 type Photo struct {
@@ -35,15 +34,6 @@ type Photos []Photo
 func (p Photos) String() string {
 	jp, _ := json.Marshal(p)
 	return string(jp)
-}
-
-// Validate gets run every time you call a "pop.Validate*" (pop.ValidateAndSave, pop.ValidateAndCreate, pop.ValidateAndUpdate) method.
-// This method is not required and may be deleted.
-func (p *Photo) Validate(tx *pop.Connection) (*validate.Errors, error) {
-	return validate.Validate(
-		&validators.StringIsPresent{Field: p.Name, Name: "Name"},
-		&validators.StringIsPresent{Field: p.ImageUrl, Name: "ImageUrl"},
-	), nil
 }
 
 // ValidateCreate gets run every time you call "pop.ValidateAndCreate" method.
