+++ models/response.go
@@ -1,4 +1,4 @@
-package model
+package models
 
 import "net/http"
