+++ service/router/home.go
@@ -3,7 +3,7 @@ package router
 import (
 	"net/http"
 
-	"github.com/NghiaTranUIT/artify-core/model"
+	"github.com/NghiaTranUIT/artify-core/models"
 	"github.com/labstack/echo"
 )
 
@@ -13,5 +13,5 @@ func (r *Router) ApplyHomeRoute(e *echo.Echo) {
 
 func (r *Router) getDefaultHome(c echo.Context) error {
 	payload := map[string]string{"Hello": "It's Artify Core"}
-	return c.JSON(http.StatusOK, model.NewSuccessReponse(payload))
+	return c.JSON(http.StatusOK, models.NewSuccessReponse(payload))
 }
