+++ service/router/photo.go
@@ -3,7 +3,7 @@ package router
 import (
 	"net/http"
 
-	"github.com/NghiaTranUIT/artify-core/model"
+	"github.com/NghiaTranUIT/artify-core/models"
 	"github.com/labstack/echo"
 )
 
@@ -19,9 +19,9 @@ func (r *Router) getFeatureToday(c echo.Context) error {
 
 	// Not found
 	if photo == nil {
-		return c.JSON(http.StatusOK, model.NewErrorResponse(err))
+		return c.JSON(http.StatusOK, models.NewErrorResponse(err))
 	}
 
 	// Success
-	return c.JSON(http.StatusOK, model.NewSuccessReponse(photo))
+	return c.JSON(http.StatusOK, models.NewSuccessReponse(photo))
 }
